cleaned up the design to use mixins for the different patterns and fixed issue in unit tests causing travis ci failures
